app/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
streamlit
fastapi
uvicorn